+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Fix CoseSigned equals
 * Base OIDs on BigInteger instead of UInt
 * Directly support UUID-based OID creation
+* Implement hash-to-curve and hash-to-scalar as per RFC9380
 
 ### 3.9.0 (Supreme 0.4.0)
 
@@ -299,7 +300,7 @@ the Tag class just cannot be directly accessed from Swift and ObjC any more.
 * Proper BIT STRING
 * BitSet (100% Kotlin BitSet implementation)
 * Recursively parsing (and encapsulating) ASN.1 structures in OCTET Strings
-* Initial pretty-printing of ASN.1 Strucutres
+* Initial pretty-printing of ASN.1 Structures
 * Massive ASN.1 builder DSL streamlining
 * More convenient explicit tagging